add teacher - separate code into files
