+++ Student.rb
@@ -1,0 +1,9 @@
+class Student < Person
+  def initialize(age, name = "unknown", parent_permission = true, classroom)
+    super(age, name , parent_permission )
+    @classroom  = classroom 
+  end
+  def play_hooky
+    puts "¯\\(ツ)/¯"
+  end
+end
