+++ Teacher.rb
@@ -1,0 +1,12 @@
+require './Person.rb'
+
+class Teacher < Person
+  def initialize(age, name = "unknown", parent_permission = true, specialization)
+    super(age, name , parent_permission )
+    @specialization  = specialization 
+  end
+
+  def can_use_services?
+    true
+  end
+end
